fix dead tree markation
If a node was marked dead on any one of its edges, we would mark
it dead, when what we want is when all of its edges are dead. This
caused tests to be cut short prematurely, especially when there was
a strict valid clause at the beginning of the buffer.